+++ CHANGELOG.md
@@ -37,3 +37,4 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Remove old sort order setting for ATLauncher packs
 - Update Flatlaf version
 - Remove server script for OSX as can share the Linux one
+- Remove servers checking tool